Hacky workaround for template replacement
By default, the template file in the role's template subdir takes
precedence over templates defined external to the role. This means that
even when use_default_template=no, with an instance_name of "haproxy"
you would end up using the template from the role.
There is no quick solution for this while keeping the role reasonably
easy to use. So we rename the default template to something unlikely to
collide with real user's expectations.